chore: rename/relocate HttpServletRequestPaths
