+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/StaticContentController.java
@@ -40,8 +40,8 @@ import static org.hisp.dhis.fileresource.FileResourceKeyUtil.makeKey;
 import static org.hisp.dhis.setting.SettingKey.USE_CUSTOM_LOGO_BANNER;
 import static org.hisp.dhis.setting.SettingKey.USE_CUSTOM_LOGO_FRONT;
 import static org.hisp.dhis.webapi.controller.StaticContentController.RESOURCE_PATH;
-import static org.hisp.dhis.webapi.service.HttpServletRequestPathParser.getContextPath;
 import static org.hisp.dhis.webapi.utils.FileResourceUtils.build;
+import static org.hisp.dhis.webapi.utils.HttpServletRequestPaths.getContextPath;
 import static org.springframework.http.HttpStatus.FOUND;
 import static org.springframework.http.HttpStatus.NO_CONTENT;
 import static org.springframework.http.HttpStatus.OK;
