+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/security/UserAccountController.java
@@ -54,7 +54,7 @@ import org.hisp.dhis.user.User;
 import org.hisp.dhis.user.UserAccountService;
 import org.hisp.dhis.user.UserService;
 import org.hisp.dhis.webapi.mvc.annotation.ApiVersion;
-import org.hisp.dhis.webapi.service.HttpServletRequestPathParser;
+import org.hisp.dhis.webapi.utils.HttpServletRequestPaths;
 import org.hisp.dhis.webmessage.WebMessageResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -121,7 +121,7 @@ public class UserAccountController {
 
     if (!userService.sendRestoreOrInviteMessage(
         user,
-        HttpServletRequestPathParser.getContextPath(request),
+        HttpServletRequestPaths.getContextPath(request),
         RestoreOptions.RECOVER_PASSWORD_OPTION)) {
       throw new ConflictException("Account could not be recovered");
     }
