+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/tracker/export/event/EventsExportController.java
@@ -69,7 +69,7 @@ import org.hisp.dhis.tracker.export.event.EventParams;
 import org.hisp.dhis.tracker.export.event.EventService;
 import org.hisp.dhis.webapi.controller.tracker.export.ChangeLogRequestParams;
 import org.hisp.dhis.webapi.controller.tracker.export.CsvService;
-import org.hisp.dhis.webapi.controller.tracker.export.JsonRequestHandler;
+import org.hisp.dhis.webapi.controller.tracker.export.FieldFilterRequestHandler;
 import org.hisp.dhis.webapi.controller.tracker.export.ResponseHeader;
 import org.hisp.dhis.webapi.controller.tracker.view.Event;
 import org.hisp.dhis.webapi.controller.tracker.view.Page;
@@ -115,7 +115,7 @@ class EventsExportController {
 
   private final EventChangeLogService eventChangeLogService;
 
-  private final JsonRequestHandler jsonRequestHandler;
+  private final FieldFilterRequestHandler fieldFilterRequestHandler;
 
   public EventsExportController(
       EventService eventService,
@@ -125,7 +125,7 @@ class EventsExportController {
       EventFieldsParamMapper eventsMapper,
       ObjectMapper objectMapper,
       EventChangeLogService eventChangeLogService,
-      JsonRequestHandler jsonRequestHandler) {
+      FieldFilterRequestHandler fieldFilterRequestHandler) {
     this.eventService = eventService;
     this.eventParamsMapper = eventParamsMapper;
     this.csvEventService = csvEventService;
@@ -133,7 +133,7 @@ class EventsExportController {
     this.eventsMapper = eventsMapper;
     this.objectMapper = objectMapper;
     this.eventChangeLogService = eventChangeLogService;
-    this.jsonRequestHandler = jsonRequestHandler;
+    this.fieldFilterRequestHandler = fieldFilterRequestHandler;
 
     assertUserOrderableFieldsAreSupported(
         "event", EventMapper.ORDERABLE_FIELDS, eventService.getOrderableFields());
@@ -321,6 +321,6 @@ class EventsExportController {
     org.hisp.dhis.tracker.export.Page<EventChangeLog> changeLogs =
         eventChangeLogService.getEventChangeLog(event, operationParams, pageParams);
 
-    return jsonRequestHandler.handle(request, "changeLogs", changeLogs, requestParams);
+    return fieldFilterRequestHandler.handle(request, "changeLogs", changeLogs, requestParams);
   }
 }
